feat(docker): use `tini` for all nodes
Don't know how it works, but it does some magic to forward shutdown signals to the correct processes, which means that our nodes shutdown quickly, instead of waiting for 10 seconds, which is the default. Hopefully this does not come with some bad side effects :fingers_crossed: At least we've had some problem with pigpio spawning multiple processes, when stopping and starting that node using docker. This could potentially help?