feat: make assignments use the start from the previous end
This means that the eel won't calculate its heading based on its current position, but instead the path of the waypoints we provide. It will also not go back to the previous waypoint, if it has skipped two. Since the heading used to be calculated from its position, the heading would point backwards. But now we can keep going forward.