chore: go back to using venv again
Apparently, we should not use `--break-system-packages`, although convenient. Besides possibly breaking stuff, ros and colcon might be sad if we accidentally install a py dep with "break system packages" that collides with any ros tool. Ros and colcon assumes some packages and paths, or something like that. But this should work well, according to
https://github.com/ros2/ros2/issues/1094